feat: airdrop command for adding transactions to network
